+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/lib/base/theme/design_system/design_system_icons.dart
@@ -48,6 +48,19 @@ class DesignSystemIcons {
   {{/has_showcase}} {{#enable_feature_otp}}
   final Icon otp = _getIcon(Icons.password);
   {{/enable_feature_otp}}
+  {{#enable_pin_code}}
+  final Icon fingerprint = _getIcon(Icons.fingerprint);
+  {{/enable_pin_code}}
+  final Icon notifications = _getIcon(Icons.notifications);
+
+  final Icon notificationsActive = _getIcon(Icons.notifications_active);
+
+  final Icon notificationsInactive = _getIcon(Icons.notifications_off);
+  {{#enable_change_language}}
+  final Icon language = _getIcon(Icons.language);
+  {{/enable_change_language}}
+  final Icon arrowForward = _getIcon(Icons.arrow_forward_ios);
+
   static Icon _getIcon(IconData iconData) => Icon(iconData);
   
 }